Passed userId to frontend
Passed userId to frontend through authService, to get it in login, autoAuthUser methods, save or remove it in localStorage.

+++ mean/mean-project/backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -67,7 +68,6 @@ router.get('', (req, res, next) => {
       maxPosts: count
     });
   });
-
 });
 
 router.put(
@@ -85,12 +85,19 @@ router.put(
     content: req.body.content,
     imagePath: imagePath
   })
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({
-      message: 'Update sucessful!',
-      postTitle: result.title,
-      postContent: result.content
-    })
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    console.log(result);
+    if (result.modifiedCount > 0) {
+      res.status(200).json({
+        message: 'Update sucessful!',
+        postTitle: result.title,
+        postContent: result.content
+      })
+    } else {
+      res.status(401).json({
+        message: 'Not authorized!'
+      })
+    }
   })
 })
 
@@ -105,10 +112,17 @@ router.get('/:id', (req, res, next)=>{
 })
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    res.status(200).json({
-      message: 'Post deleted!'
-    });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId }).then(result => {
+    console.log(result);
+    if (result.deletedCount > 0) {
+      res.status(200).json({
+        message: 'Delete sucessful!'
+      })
+    } else {
+      res.status(401).json({
+        message: 'Not authorized!'
+      })
+    }
   }).catch()
 
 })
